kindle_vocab/requirements.txt
flask
python-dotenv
requests
stanza
verbecc